Add smoke test and harness
The run test file creates an extrernal network and halts execution of the test container until the stack is up.

Responses and errors are logged.

Tests will access the docker, patroni apis and optionally the dbs directly through haproxy.

This intially test just checks to see if the stack is up and ready